Move some url logic in separate utils class
