+++ app/src/main/java/com/duckduckgo/mobile/android/util/DDGUtils.java
@@ -276,38 +276,6 @@ public final class DDGUtils {
 				}
 			}
 		}
-	  	  
-		/**
-		 * Checks to see if URL is DuckDuckGo SERP
-		 * Returns the query if it's a SERP, otherwise null
-		 * 
-		 * @param url
-		 * @return
-		 */
-		static public String getQueryIfSerp(String url) {
-			if(!isSerpUrl(url)) {
-                return null;
-            }
-			
-			Uri uri = Uri.parse(url);
-			String query = uri.getQueryParameter("q");
-			if(query != null)
-				return query;
-			
-			String lastPath = uri.getLastPathSegment();
-			if(lastPath == null)
-				return null;
-			
-			if(!lastPath.contains(".html")) {
-				return lastPath.replace("_", " ");
-			}
-			
-			return null;
-		}
-
-    public static boolean isSerpUrl(String url) {
-        return url.contains("duckduckgo.com");
-    }
 		
 	private static boolean isIntentSafe(Context context, Intent intent) {
 		// Verify it resolves
@@ -348,30 +316,11 @@ public final class DDGUtils {
 	}
 
     public static void searchExternal(Context context, String term) {
-		String url = DDGControlVar.mDuckDuckGoContainer.torIntegration.isTorSettingEnabled() ? DDGConstants.SEARCH_URL_ONION : DDGConstants.SEARCH_URL;
-		if (DDGControlVar.regionString.equals("wt-wt")) {    // default
-			url = url.replace("ko=-1&", "") + URLEncoder.encode(term);
-		}
-		else {
-			url = url.replace("ko=-1&", "") + URLEncoder.encode(term) + "&kl=" + URLEncoder.encode(DDGControlVar.regionString);
-		}
+		String url = UrlUtils.getSearchUrlForterm(
+				DDGControlVar.mDuckDuckGoContainer.torIntegration.isTorSettingEnabled() ? UrlUtils.SEARCH_URL_ONION : UrlUtils.SEARCH_URL,
+				term);
         Intent browserIntent = new Intent(Intent.ACTION_VIEW, Uri.parse(url));
         execIntentIfSafe(context, browserIntent);
-    }
-
-    public static String getUrlToDisplay(String url) {
-        if(url==null || url.length()==0) {
-            return "";
-        }
-        if (url.startsWith("https://")) {
-            url = url.replace("https://", "");
-        } else if (url.startsWith("http://")) {
-            url = url.replace("http://", "");
-        }
-        if (url.startsWith("www.")) {
-            url = url.replace("www.", "");
-        }
-        return url;
     }
 
     public static SCREEN getScreenByTag(String tag) {
