+++ app/src/main/res/menu/web.xml
@@ -15,12 +15,6 @@
     <item android:id="@+id/action_external"
           android:title="@string/action_view_external"
           android:orderInCategory="50"
-          app:showAsAction="never" />
-
-    <item android:id="@+id/action_remove_recent"
-          android:title="Remove"
-          android:visible="false"
-          android:orderInCategory="60"
-          app:showAsAction="never"/>-->
+          app:showAsAction="never" />-->
 
 </menu>
